continue serializing map functions
