+++ ensembl_map/exceptions.py
@@ -1,3 +1,12 @@
+class ConvertError(Exception):
+    def __init__(self, from_type, to_type):
+        self.from_type = from_type
+        self.to_type = to_type
+
+    def __str__(self):
+        return f"Cannot convert {self.from_type} directly to {self.to_type}"
+
+
 class OutsideCdsError(Exception):
     def __init__(self, tobj, position):
         self.tobj = tobj
