+++ ensembl_map/util.py
@@ -1,6 +1,16 @@
 import re
 
 
+def assert_valid_position(start=None, end=None):
+    if start is not None and start < 1:
+        raise ValueError(f"Start must be >= 1 ({start})")
+    if end is not None and end < 1:
+        raise ValueError(f"End must be >= 1 ({end})")
+
+    if (start is not None and end is not None) and start > end:
+        raise ValueError(f"Start ({start}) must be <= end ({end})")
+
+
 def is_ensembl_id(feature):
     """The given string looks like and Ensembl ID."""
     return bool(re.match(r"ENS[A-Z]\d{8}", feature))
